nix: add preliminary docker image scaffold
Summary: These certainly don't work yet -- the image specification probably
needs work and the startup scripts need to be ported -- but this gives us
something to work with for demos and tests and other "external world" use cases.

This changes the layout of the derivation names so now things are more nested; now
we have four derivations, renaming the original ones:

    psql_14 ----> psql_14/bin
	psql_15 ----> psql_15/bin

And there's now a `psql_14/docker` (c.f. psql_15) as well.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Id4340d4f8747d6adead7cb52193687d3