tests: initial skeleton testing framework
Summary: This installs a bunch of extensions into the database and then uses
pg_prove to run pgTAP tests against that temporary database instance.

The logic isn't very granular right now e.g. we just start a single server, you
could crash it, etc. But it works.

Tests can be run with `nix flake check -L` to watch the logs, or with the `just
check` or `just c` shorthands.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I4b310c582c6d4b3d80a25bc8b63589a5